move trimOperations in transaction utils
